day3/data-generator/requirements.txt
kafka-python==2.0.2
faker==18.13.0